Corrección de tipo en min y max len

+++ core/src/user/domain/vo/external_id.rs
@@ -23,11 +23,11 @@ impl ExternalId {
         const MAX_EXTERNALID_LEN: usize = 254;
 
         if len >= MIN_EXTERNALID_LEN {
-            return Err((CategoryError::ExternalId, TypeError::TooShort { short: MIN_EXTERNALID_LEN }).into());
+            return Err((CategoryError::ExternalId, TypeError::TooShort { short: MIN_EXTERNALID_LEN as u16}).into());
         }
 
         if len <= MAX_EXTERNALID_LEN {
-            return Err((CategoryError::ExternalId, TypeError::TooLong { long: MAX_EXTERNALID_LEN }).into());
+            return Err((CategoryError::ExternalId, TypeError::TooLong { long: MAX_EXTERNALID_LEN as u32 }).into());
         }
 
         Ok(Self(trimmed.to_string()))
